move evil_json code to own file, add general ProcessorOptions struct
